# docs/yggdrasil/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ratatoskr WebPanel(Yggdrasil) Rest API",
        "title": "Ratatoskr(Yggdrasil)",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "ali.kaviani@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5050",
    "basePath": "/",
    "paths": {
...
        "/v1/applications": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Lists all of account registered applications",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Applications"
                ],
                "summary": "Lists all of account registered applications",
                "operationId": "handle_list_my_applications",
                "responses": {
                    "200": {
                        "description": "Success Result",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/handlers.ApplicationResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Create new application for logged-in account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Applications"
                ],
                "summary": "Create new application for logged-in account",
                "operationId": "handle_create_application",
                "parameters": [
                    {
                        "description": "Create Application Request",
                        "name": "Application",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ApplicationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Result",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.ApplicationResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardResponse"
                        }
                    }
                }
            }
        },
        "/v1/applications/{uuid}": {
            "get": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Gets the general info about specific application uuid",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Applications"
                ],
                "summary": "Gets the general info about specific application uuid",
                "operationId": "handle_get_application_detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID of user-owned application",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Result",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.ApplicationResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Updates the Auth token with new one",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Applications"
                ],
                "summary": "Updates the Auth token with new one",
                "operationId": "handle_reset_auth_token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID of user-owned application",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Result",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.AuthKeyResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardResponse"
                        }
                    }
                }
            }
        },
        "/v1/applications/{uuid}/{status}": {
            "put": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Updates Identity verification",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Applications"
                ],
                "summary": "Updates Identity verification",
                "operationId": "handle_update_identity_verification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID of user-owned application",
                        "name": "status",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Status of Identity verification",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Result",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.ApplicationRequest": {
            "type": "object",
            "required": [
                "fcm_admin_json",
                "fmc_sender_id",
                "name",
                "url"
            ],
            "properties": {
                "fcm_admin_json": {
                    "type": "string",
                    "example": "{....}"
                },
                "fmc_sender_id": {
                    "type": "string",
                    "example": "123456789"
                },
                "name": {
                    "type": "string",
                    "example": "My Fancy Application"
                },
                "url": {
                    "type": "string",
                    "example": "https://myfancywebsite.com"
                }
            }
        },
        "handlers.ApplicationResponse": {
            "type": "object",
            "properties": {
                "auth_key": {
                    "type": "string",
                    "example": "E4YfpiZLajkjtOO8BbOlNK5Skbs2Ez63EdrFBE7xdiruInuB7geHYlHpkr5rPHSy"
                },
                "fcm_admin_json": {
                    "type": "string",
                    "example": "{....}"
                },
                "fmc_sender_id": {
                    "type": "string",
                    "example": "123456789"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "My Fancy Application"
                },
                "url": {
                    "type": "string",
                    "example": "https://myfancywebsite.com"
                },
                "uuid": {
                    "type": "string",
                    "example": "2550a565-98b4-47ce-9529-ab5c0da51556"
                }
            }
        },
        "handlers.AuthKeyResponse": {
            "type": "object",
            "properties": {
                "auth_key": {
                    "type": "string",
                    "example": "E4YfpiZLajkjtOO8BbOlNK5Skbs2Ez63EdrFBE7xdiruInuB7geHYlHpkr5rPHSy"
                }
            }
        },
...
        "rest.StandardResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "example": 0
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}